requirements.txt
requests>=2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
openai
pyyaml
loguru